+++ bindings/nodejs/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Rename `Client::listen` to `listenMqtt`, `Client::clearListeners` to `clearMqttListeners`;
+- Moved `minimumRequiredStorageDeposit()` from `Account` to `Client`;
 
 ## 1.0.0-rc.2 - 2023-07-05